Refactor:#132 방과 게임의 이벤트 핸들러 분리 및 SecurityUtil 메서드를 static 메서드로 수정
