+++ src/main/java/mutsa/yewon/talksparkbe/global/util/SecurityUtil.java
@@ -6,11 +6,10 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 
-@Component
 @Log4j2
 public class SecurityUtil {
 
-    public Long getLoggedInUserId() {
+    public static Long getLoggedInUserId() {
 
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
